requirements.txt
aiohttp==3.9.4
disnake==2.9.2
pillow==10.3.0
prisma==0.13.1
